frontend/requirements.txt
streamlit
pandas
plotly
requests
openpyxl
reportlab
pillow
